Move read functions to utils module and read_rgb by default
